useStoreとcreateCollectionStoreは抽象化できたので切り出した
